pyproject.toml
[tool.poetry]
name = "research-ai"
version = "0.1.0"
description = "Modular research ingestion and metadata platform"
authors = ["Smarth Bakshi <bakshismarth.20@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
  { include = "researchai_core", from = "core/src" },
  { include = "api", from = "api/src" },
  { include = "services", from = "services/" },
  { include = "scripts", from = "scripts/" },
  { include = "tests", from = "tests/" }
]

[tool.poetry.dependencies]
flask = "2.2.5"
connexion = "<3.0"
werkzeug = "<3.0"
python = ">=3.10,<3.13"
requests = "^2.31.0"
tenacity = "^8.2.3"
pydantic = "^2.7.1"
services = "^0.1.1"
boto3 = "^1.40.11"
python-dotenv = "^1.1.1"
apache-airflow = "^3.0.4"
pypdf = "^6.0.0"
pytesseract = "^0.3.13"
pdf2image = "^1.17.0"
minio = "^7.2.16"
pdfminer-six = "^20250506"
sentence-transformers = "^3.3.1"
opensearch-py = "^2.7.1"
sqlalchemy = "^1.4.0,<2.0"
pillow = "^11.0.0"
numpy = "^1.23.0,<2.0"


[tool.poetry.group.dev.dependencies]
black = "^24.0"
pytest = "^8.2"
ruff = "^0.4.0"
mypy = "^1.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]